requirements.txt
torch-geometric==1.7.2
numpy==1.19.5
pandas==1.3.5
torch-scatter==2.0.6
torch-sparse==0.6.8
torch-cluster==1.5.8
torch-spline-conv==1.2.1